Line creation handle
